unit test for basic oscillator passes

+++ scala-synth-main/src/main/scala/org/kylerow/scalasynth/SSConfiguration.scala
@@ -1,6 +1,6 @@
 package org.kylerow.scalasynth
 
 class SSConfiguration {
-	def getSampleRate() :Double = 44100d;
+	def getSampleRate() :Int = 44100;
 	def getBufferSize() :Int = 4096;
 }
